Added merging GA output with loaded dataset
Implemented a method that returns a datasaet without last column (class index)
Implemented a method that merges output from the genetic algorithm (bitstring that represents prediction of last column) with the dataset without last column, so that the new dataset could be sent into training of Naive Bayes model.
Added stopping criteria for maximum number of generations in genetic algorithm.